move funs of framework/deployment to e2e/apps
